:package: Refactor UUIDConsumer
