+++ src/config/config.cpp
@@ -19,8 +19,8 @@
 
 #include "config.hpp"
 
-#include "../UUIDConsumer/UUIDConsumerKafka.h"
 #include "../UUIDProducer/UUIDKafkaTopicProducer.h"
+#include "../uuid_consumer/kafka_uuid_consumer.hpp"
 
 #include <rapidjson/document.h>
 #include <rapidjson/error/en.h>
@@ -72,7 +72,7 @@ public:
   KafkaUUIDConsumerFactory &operator=(KafkaUUIDConsumerFactory &&) = delete;
   ~KafkaUUIDConsumerFactory() {}
 
-  std::unique_ptr<UUIDConsumer> create() {
+  std::unique_ptr<UUIDConsumer::UUIDConsumer> create() {
     string errstr;
     unique_ptr<RdKafka::Conf> conf(
         RdKafka::Conf::create(RdKafka::Conf::CONF_GLOBAL)),
@@ -82,8 +82,9 @@ public:
     rdkafka_set_conf_vector(this->m_kafka_consumer_conf, conf.get(), "kafka");
     conf->set("default_topic_conf", tconf.get(), errstr);
 
-    return std::unique_ptr<UUIDConsumer>(new CounterUUIDJSONKafkaConsumer(
-        this->m_read_topics, m_json_uuid_key, conf.get()));
+    return std::unique_ptr<UUIDConsumer::UUIDConsumer>(
+        new UUIDConsumer::CounterUUIDJSONKafkaConsumer(
+            this->m_read_topics, m_json_uuid_key, conf.get()));
   }
 
 private:
